requirements.txt
dbt-bigquery
google-cloud-bigquery